update ubuntu to 20.0.4

+++ bootstrap.sh
@@ -1,20 +1,23 @@
 #!/bin/bash
 
 if [ -f /etc/redhat-release ] ;  then
-    sudo yum install -y vim-enhanced mc tree git > /dev/null
+    sudo yum install -y vim-enhanced tree git > /dev/null
     vagrant_user="vagrant"
 fi
 
 if [ -f /etc/lsb-release ] ; then
-    echo "ubuntu"
+    #echo "ubuntu"
     grep -q "archive.ubuntu.com" /etc/apt/sources.list && sed -i.bak -e 's/archive.ubuntu.com/ftp.daum.net/g' -e 's/security.ubuntu.com/ftp.daum.net/g' /etc/apt/sources.list && sudo apt update
-    sudo apt install -y mc tree git
-    vagrant_user="ubuntu"
+    sudo apt install -y git > /dev/null
+
+    vagrant_user="vagrant"
 
     date | grep -q UTC && sudo timedatectl set-timezone Asia/Seoul
+    timedatectl | egrep "synchronized:yes"  || sudo timedatectl set-ntp on
 fi
 
-cat > /home/$vagrant_user/.vimrc <<EOF
+if ! [ -f /home/$vagrant_user/.vimrc ] ; then
+    cat > /home/$vagrant_user/.vimrc <<EOF
 set expandtab
 " set cindent
 set tabstop=4
@@ -23,7 +26,9 @@ set shiftwidth=4
 set autoindent
 EOF
 
-chown $vagrant_user.$vagrant_user /home/$vagrant_user/.vimrc
+    chown $vagrant_user.$vagrant_user /home/$vagrant_user/.vimrc
+
+fi
 
 if ! [ -f /etc/profile.d/change_ps1.sh ] ; then
     cat > /etc/profile.d/change_ps1.sh <<'EOF'
